Further optimized rendering speed and preview generation

+++ Mandelbrot.Domain.Interface/Calculation/Algorithms/IPointBasedFractal.cs
@@ -2,6 +2,6 @@ namespace Mandelbrot.Domain.Calculation.Algorithms
 {
 	public interface IPointBasedFractal : IFractalAlgorithm
 	{
-		decimal CalculateSinglePoint(int x, int y, IFractalSettings settings);
+		double CalculateSinglePoint(int x, int y, IFractalSettings settings);
 	}
 }
